requirements.txt
Flask==3.0.0
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
pycryptodome==3.20.0
matplotlib==3.8.2
